requirements.txt
pandas==2.2.3
fastapi[standard]==0.115.11
dotenv==0.9.9
psycopg2-binary==2.9.10